Changed ints to size_ts

+++ intro/threads.c
@@ -11,12 +11,12 @@
 #include <semaphore.h>
 #endif
 
-volatile int counter = 0; 
-int loops;
+volatile size_t counter = 0; 
+size_t loops;
 
 void *worker(void *arg) 
 {
-    int i;
+    size_t i;
     for (i = 0; i < loops; i++) 
     {
         counter++;
@@ -33,12 +33,12 @@ int main(int argc, char *argv[])
     } 
     loops = atoi(argv[1]);
     pthread_t p1, p2;
-    printf("Initial value : %d\n", counter);
+    printf("Initial value : %zu\n", counter);
     assert(pthread_create(&p1, NULL, worker, NULL) == 0);
     assert(pthread_create(&p2, NULL, worker, NULL) == 0);
     assert(pthread_join(p1, NULL) == 0);
     assert(pthread_join(p2, NULL) == 0);
-    printf("Final value   : %d\n", counter);
+    printf("Final value   : %zu\n", counter);
     return 0;
 }
